Moved cname and zone change logic to frontend subscriber
