+++ inc/Engine/CDN/RocketCDN/DataManagerSubscriber.php
@@ -86,8 +86,6 @@ class DataManagerSubscriber implements Subscriber_Interface {
 			'wp_ajax_rocketcdn_validate_token_cname' => 'validate_token_cname',
 			self::CRON_EVENT                         => 'maybe_disable_cdn',
 			'wp_rocket_upgrade'                      => [ 'refresh_cdn_cname', 10, 2 ],
-			'pre_get_rocket_option_cdn_cnames'       => [ 'set_cdn_cnames', 9 ],
-			'pre_get_rocket_option_cdn_zone'         => [ 'set_cdn_zone', 9 ],
 		];
 	}
 
@@ -526,65 +524,5 @@ class DataManagerSubscriber implements Subscriber_Interface {
 		$this->options->set( 'cdn_cnames', $cdn_cnames );
 
 		$this->options_api->set( 'settings', $this->options->get_options() );
-	}
-
-	/**
-	 * Sets the CDN CNAME from the RocketCDN subscription data on the filter level.
-	 *
-	 * @since 3.22
-	 *
-	 * @param mixed $value The current pre-filter value.
-	 *
-	 * @return mixed The CDN CNAME array if RocketCDN is active, or the original value.
-	 */
-	public function set_cdn_cnames( $value ) {
-		$cdn_url = $this->get_rocketcdn_url();
-
-		if ( empty( $cdn_url ) ) {
-			return $value;
-		}
-
-		return [ $cdn_url ];
-	}
-
-	/**
-	 * Sets the CDN zone from the RocketCDN subscription data on the filter level.
-	 *
-	 * @since 3.22
-	 *
-	 * @param mixed $value The current pre-filter value.
-	 *
-	 * @return mixed The CDN zone array if RocketCDN is active, or the original value.
-	 */
-	public function set_cdn_zone( $value ) {
-		$cdn_url = $this->get_rocketcdn_url();
-
-		if ( empty( $cdn_url ) ) {
-			return $value;
-		}
-
-		return [ 'all' ];
-	}
-
-	/**
-	 * Gets the CDN URL from the RocketCDN subscription data.
-	 *
-	 * @since 3.22
-	 *
-	 * @return string The CDN URL if subscription is active, empty string otherwise.
-	 */
-	private function get_rocketcdn_url(): string {
-		if ( 'rocketcdn' !== $this->options->get( 'cdn_type' ) ) {
-			return '';
-		}
-
-		// RFT Todo: Might require a different approach during the API integration to check for subscription status.
-		$subscription = $this->api_client->get_subscription_data();
-
-		if ( 'running' !== $subscription['subscription_status'] ) {
-			return '';
-		}
-
-		return $subscription['cdn_url'];
 	}
 }
