+++ inc/Engine/CDN/RocketCDN/ServiceProvider.php
@@ -23,6 +23,7 @@ class ServiceProvider extends AbstractServiceProvider {
 		'rocketcdn_api_client',
 		'rocketcdn_options_manager',
 		'rocketcdn_data_manager_subscriber',
+		'rocketcdn_frontend_subscriber',
 		'rocketcdn_rest',
 		'rocketcdn_rest_subscriber',
 		'rocketcdn_admin_subscriber',
@@ -69,6 +70,14 @@ class ServiceProvider extends AbstractServiceProvider {
 					'options',
 					'options_api',
 					'user_client',
+				]
+			);
+		// RocketCDN Frontend subscriber.
+		$this->getContainer()->addShared( 'rocketcdn_frontend_subscriber', FrontendSubscriber::class )
+			->addArguments(
+				[
+					'options',
+					'rocketcdn_api_client',
 				]
 			);
 		// RocketCDN REST API pages controller.
